refactor: styledComponents 분리
