+++ client/src/components/ProductCard/ProductCard.style.ts
@@ -1,0 +1,46 @@
+import styled from 'styled-components';
+
+export const ProductContainer = styled.ul`
+  margin-top: 1rem;
+  margin-bottom: 7rem;
+  width: 100%;
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(320px, 1fr));
+  grid-gap:10px; 
+  justify-items: center; 
+  align-items: center; 
+  grid-row-gap: 3rem;
+`
+
+export const ProductImage = styled.img`
+  height: 100%;
+  width: 100%;
+  object-fit: cover;
+  object-position: center;
+  cursor: pointer;
+`
+export const HoverOverlay = styled.div`
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(255, 255, 255, 0.2);
+  opacity: 0;
+  transition: opacity 0.5s;
+  pointer-events: none;
+  li:hover & {
+    opacity: 1;
+  }
+`;
+
+export const ModalWrapper = styled.div`
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`;
