+++ client/src/pages/store/StoreDetail.tsx
@@ -1,9 +1,16 @@
-import styled from 'styled-components';
 import axios from 'axios';
 import { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
 import { BiArrowBack} from 'react-icons/bi';
 import ProductCard from '../../components/ProductCard/ProductCard';
+import {
+  StoreProductSection,
+  ProductListTitle,
+  StoreDetails,
+  StoreDetailSection,
+  DetailInfo,
+  Detail
+} from './StoreDetail.style';
 
 interface Product {
   productId: number;
@@ -158,7 +165,7 @@ const data: Store =
     <>
         <StoreDetailSection>     
           <Link to='/store'>
-            <BiArrowBack style={{fontSize:'40px', color: 'var(--dark-purple)', marginTop:'11rem', position: 'absolute', left:'12rem', cursor:'pointer'}}/>
+            <BiArrowBack style={{fontSize:'40px', color: 'var(--dark-purple)', marginTop:'11rem', position: 'absolute', left:'11rem', cursor:'pointer'}}/>
           </Link>     
             <StoreDetails>            
                 <div style={{backgroundImage:`url('${data.storeImage}')`, width: '280px', height: '280px', borderRadius: '50%', backgroundSize:'cover'}}></div>
@@ -175,61 +182,14 @@ const data: Store =
         </StoreDetailSection>
         <StoreProductSection>
             <div style={{width: '80%'}}>
-                <ProductTitle>Custom</ProductTitle>
+                <ProductListTitle>Custom</ProductListTitle>
                 <ProductCard data={data.customProductInfoList} storeId={data.storeId} storeName={data.storeName}/>                
-                <ProductTitle>We Made It</ProductTitle>
+                <ProductListTitle>We Made It</ProductListTitle>
                 <ProductCard data={data.standardProductInfoList} storeId={data.storeId} storeName={data.storeName}/>
             </div>
         </StoreProductSection>
     </>
   );
 }
-
-const StoreProductSection = styled.section`
-display: flex;
-flex-direction: column;
-align-items: center;
-`
-const ProductTitle = styled.h4`
-color: var(--dark-gold);
-font-weight: 700;
-`
-const StoreDetails = styled.section`
-margin-top: 14rem;
-display: flex;
-justify-content: center;
-align-items: center;
-margin-bottom: 5rem;
-`
-
-const StoreDetailSection = styled.section`
-margin-bottom: 4rem;
-display: flex;
-flex-direction: column;
-align-items: center;
-    &::after {
-    content:"";
-    display: block;
-    background-color: var(--light-gray);
-    height: 1.4px;
-    width: 80%;
-    margin-top: 20px;
-  }
-`
-
-const DetailInfo = styled.p`
-    color: var(--bright-black);
-    font-weight: 400;
-    font-size: 13px;
-    margin-top: 10px;
-    margin-bottom: 20px;
-`
-
-const Detail = styled.span`
-    color: var(--light-gray);
-    font-size: 13px;
-    margin-right: 5px;
-    font-weight: 600;
-`
 
 export default StoreDetail;
